Update README.md
Comment 2:

"The details of the preprocessing of the data (e.g., removal of duplicate transactions, three-tier tracking of wallets) are explained in 'README.md,' but the specific code and parameters (e.g., filtering conditions, cluster threshold) are not described. Please add reference values for 'MAX_DEPTH:' and 'TX_COUNT_THRESHOLD:'. Also, mention that these are listed in the 'README.md' file."